Homework 25 (29) | unfinished 2

+++ KotlinBasic/25_29_ContentProvider/app/src/main/java/ru/skillbox/a25_29_contentprovider/contact_detail/BaseContactInfoHolder.kt
@@ -5,10 +5,10 @@ import androidx.recyclerview.widget.RecyclerView
 
 abstract class BaseContactInfoHolder(view: View) : RecyclerView.ViewHolder(view) {
 
-    private var currentId: Long? = null
+    private var currentId: Int? = null
 
     protected fun bindMainInfo(
-        id: Long
+        id: Int
     ) {
         currentId = id
     }
